Make import code from same python installation
This addresses some problems with the executable script:

  - It is starting the new code with 'python3', which is not the same as
  /usr/bin/python3 in general, for example when the user is using
  anaconda or virtualenv.

  - Launching a second executable makes the code unfriendly to tools
  such as debuggers.

  - The return code is not propagated.

A minimal change, implemented here, is to wrap the runnable code of the
library module in a main function (similar to the recommended approach
of e.g. setuptools) and import it from the code in the bin folder
(following the approach by tools such as flit).