all: Ensure a unique service name, and handle collisions gracefully internally and in the ui.
A preferred service name will be checked if it already exists, and
a [n] will be appended to the hostname part of the service name.

The actual tracking internally will be based only on the true hostname
and ip address.  The service name will be kept to match against a
subsequent remove_service() call.  This allows a remote to go offline
as hostname, return as hostname[1] and still be recognized as hostname.

This should resolve the service name collision issue in #22 when hostnames
are duplicated.