Increase discovery flush time, and extend the initialization view for a longer time to account for it.
Depending on network topology (multiple wireless access points, for
instance), the zeroconf service browser can fail to detect changes in
a remote machine's state, especially if that remote did not cleanly
shut down (and perform its own de-registration).

For this reason a 'dummy' ServiceInfo was created and immediately
removed in order to resolve the inaccurate state that other browsers
were in.  This works fine, but is not 100% reliable.  Forcing a longer
delay between these steps seems to improve success.

If this continues to be an issue, we could continually poll the last-
known ip address of a remote machine to check for it coming back online,
bypassing zeroconf for these re-connections.