Transfers: Allow compression when both sides have support and have it enabled.
This results in a substantial reduction in transfer times in most cases
where the file(s) being transfered aren't already compressed.

In rough measurements, I went from an average of 10-12mb/s to around
30mb/s on both large single files and large amounts of smaller files.

In certain situations this can actually reduce performance quite a bit,
which I noticed in transferring files from a host to a vbox guest using
a bridged connection. This might occur in other situations as well, so
it can be disabled if necessary.

Both ends of the transfer must support and enable compression for it
to work, otherwise it transfers normally.

This is implemented in such a way that once (and if) native grpc
compression becomes available, it can be removed easily.