prefs.py: Fix typo in port dupe guard.
This code should have been preventing a situation where both data
and reg ports were the same at startup. The gui prevents the user
from doing this, but there are a couple of corner cases where this
can occur on its own - namely, during an upgrade, when the original
port is set to 42001 - the reg port will also be set at that value,
which is its default.

Fixes #107