Transfers: Make transfers cancellable by the receiving computer once more.
2cd3b660c added client- and server-side interceptors to handle
compression mid-flight. This, unfortunately, caused the receiving
machine to lose access to the original response's .cancel() method,
as the interceptor was now passing a generator function.

Now, the interceptor will return both the generator and the original
response, with the response being used solely as a 'cancellable'
object. This is very undesirable because that means StartTransfer is
no longer returning what its protobuf definition claims it should,
but it works until/unless I can figure out a 'correct' solution.