Fix network device name fetching
lshw messes up its json output to not supply a valid json format for multi-network-device configs. For single-network-device machines, the json itself is valid but does not represent an array.
This fixes both cases for our use case.
Ubuntu ships with a patched version of lshw so this problem does not happen on Ubuntu or Mint. It does however happen on Fedora 35.