transfers: Always remove existing files before beginning a transfer.
The user must approve overwriting, or they've must have allowed it
via preferences. Either way, if we've made it this far, we can get
rid of existing files before starting the transfer.

This eliminates the risk of any existing symbolic directory links
potentially being followed outside of the download directory, which
would not be detected by 5244c33d4c109.