remote_registration: Do not loop registration attempts.
Zeroconf uses a single thread for its ServiceBrowser. With multiple
machines are on a network, having one with a different code was
potentially preventing the remaining machines from registering with
one another, depending upon the order of their discovery.

Previous behavior:
- discovery machine 1
-- start registration loop
---- group code check, continue on success or loop indefinitely,
     blocking any remaining machines from being discovered.
-- complete loop
- continue setup of remote....
- discovery machine 2
...

New behavior:
- discovery machine 1
-- attempt to register once.
- if failure, ignore the remote, otherwise continue setup
- discovery machine 2
...

This issue affected both versions of registration (udp and grpc-
based) so seems to have existed for some time.

ref: #157