Remove restrictions on having a save directory in /run/*
There are a lot of different scenarios in the wild where this is
interfering with using Warpinator at all, and considering how
permissions are set in /run, there's not much danger here.

Landlock is becoming and hopefully will be universally enabled
by default on distro kernels. Also the transfer symlink attack
potential is already low because of other checks that were
added in a4a0b0dfcc81.