src/transfers.py: Follow toplevel symbolic links when gathering files to send.
Previous behavior was to never follow links, under any conditions.

Now if the file selection contains a symbolic link, the link will
be followed, and the linked-to files will be sent in place of
the link.

This does not alter any existing behavior with links inside of
folders - these will never be followed, and the link itself will
be sent instead.

Fixes: #212